requirements.txt
beautifulsoup4
joblib==1.0.1
torchmetrics==0.4.1
stanza==1.2.2
more_itertools==8.8.0
tqdm==4.61.2
pytorch_lightning==1.3.8
nltk==3.6.2
requests==2.26.0
numpy==1.21.1
six~=1.16.0
h5py~=3.3.0
d2l==1.0.0a1
torchvision